[SpringBootPractice][Marcus Luk]Test: Added test case for exceptions in service test.

+++ src/main/java/com/thoughtworks/springbootemployee/service/CompanyService.java
@@ -58,7 +58,11 @@ public class CompanyService {
         throw new CompanyNotFoundException();
     }
 
-    public void deleteCompany(String companyId) {
-        companyRepository.deleteById(companyId);
+    public void deleteCompany(String companyId) throws CompanyNotFoundException {
+        if (companyRepository.existsById(companyId)) {
+            companyRepository.deleteById(companyId);
+            return;
+        }
+        throw new CompanyNotFoundException();
     }
 }
